+++ src/utils.ts
@@ -1,4 +1,5 @@
 import { window, commands } from 'vscode';
+import * as fs from 'fs-extra';
 
 export async function showOpenSettingsPrompt(
     errorMessage: string
@@ -12,6 +13,18 @@ export async function showOpenSettingsPrompt(
     }
 }
 
+export async function showReportIssueErrorMessage(
+    errorMessage: string
+): Promise<void> {
+    const selected = await window.showErrorMessage(
+        errorMessage,
+        'Report Issue'
+    );
+    if (selected === 'Report Issue') {
+        await commands.executeCommand('psalm.reportIssue');
+    }
+}
+
 export async function showErrorMessage(
     errorMessage: string
 ): Promise<string | undefined> {
@@ -22,4 +35,12 @@ export async function showWarningMessage(
     errorMessage: string
 ): Promise<string | undefined> {
     return window.showWarningMessage(errorMessage);
+}
+
+export async function formatFromTemplate(templatePath: string, ...args: any[]) {
+    const template = await fs.readFile(templatePath, 'utf8');
+
+    return template.replace(/{(\d+)}/g, (match, number) =>
+        args[number] === undefined ? match : args[number]
+    );
 }